// src/app/models/user.ts
export interface userList {
    CST: boolean;
    Client_ID: number;
    VIP: boolean;
    active: boolean;
    allTickets: boolean;
    canApproveChanges: boolean;
    emailAddress: string;
    isChangeApproval: boolean;
    isMailOn: boolean;
    isUserAdmin: boolean;
    mobile: string,
    name: string,
    primarySiteContact: boolean,
    spend: boolean,
    telephone: string,
    title: boolean,
    totalCount: number,
    ultraVIP: boolean,
    userId: number,
    userType: string,
    userTypeId: number;
}